% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lagr.dispatch.r
\name{lagr.dispatch}
\alias{lagr.dispatch}
\title{Dispatch the fitting of local models to parallel cores, if registered}
\usage{
lagr.dispatch(x, y, family, coords, fit.loc, oracle, D, bw, bw.type, verbose,
  varselect.method, prior.weights, tuning, predict, simulation, kernel, min.bw,
  max.bw, min.dist, max.dist, tol.loc, lambda.min.ratio, n.lambda,
  lagr.convergence.tol, lagr.max.iter, jacknife = FALSE,
  bootstrap.index = NULL)
}
\arguments{
\item{x}{matrix of observed covariates}

\item{y}{vector of observed responses}

\item{family}{exponential family distribution of the response}

\item{coords}{matrix of locations, with each row giving the location at which the corresponding row of data was observed}

\item{fit.loc}{matrix of locations where the local models should be fitted}

\item{D}{pre-specified matrix of distances between locations}

\item{bw}{bandwidth parameter}

\item{bw.type}{type of bandwidth - options are \code{dist} for distance (the default), \code{knn} for nearest neighbors (bandwidth a proportion of \code{n}), and \code{nen} for nearest effective neighbors (bandwidth a proportion of the sum of squared residuals from a global model)}

\item{verbose}{print detailed information about our progress?}

\item{varselect.method}{criterion to minimize in the regularization step of fitting local models - options are \code{AIC}, \code{AICc}, \code{BIC}, \code{GCV}}

\item{tuning}{logical indicating whether this model will be used to tune the bandwidth, in which case only the tuning criteria are returned}

\item{kernel}{kernel function for generating the local observation weights}

\item{tol.loc}{tolerance for the tuning of an adaptive bandwidth (e.g. \code{knn} or \code{nen})}
}
\description{
Loops through estimation locations with foreach, sending each local model to a core for fitting. If zero or one cores are registered, then foreach computes the local models sequentially.
}

