% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lagr.parse.model.frame.r
\name{lagr.parse.model.frame}
\alias{lagr.parse.model.frame}
\title{This function takes input to the lagr or lagr.tune function, returning a model matrix, coordinates, weights, and response.}
\usage{
lagr.parse.model.frame(formula, data, family, weights, coords, fit.loc, longlat,
  na.action, contrasts)
}
\description{
This function takes input to the lagr or lagr.tune function, returning a model matrix, coordinates, weights, and response.
}

