\name{SGL}
\alias{SGL}
\title{Do the local adaptive grouped regularization}
\usage{
SGL(data, index, weights = NULL, type = "linear", maxit = 1000,
  thresh = 0.001, min.frac = 0.1, nlam = 20, delta = 2, gamma = 0.8,
  verbose = FALSE, step = 1, reset = 10, lambdas = NULL,
  unpenalized = NULL)
}
\description{
Do the local adaptive grouped regularization
}

