% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grouplasso}
\alias{grouplasso}
\title{Do the local adaptive grouped regularization}
\usage{
grouplasso(data, index, weights = NULL, type = "linear", maxit = 1000,
  thresh = 0.001, min.frac = 0.1, nlam = 20, delta = 2, gamma = 0.8,
  verbose = FALSE, momentum = 1, reset = 10, lambda = NULL,
  unpenalized = NULL)
}
\description{
Do the local adaptive grouped regularization
}

