% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grouplassoRun.r
\name{grouplassoRun}
\alias{grouplassoRun}
\title{Get the sequence of tuning parameters and then make the call to the C code that implements the adaptive group lasso}
\usage{
grouplassoRun(data, index, weights, adaweights, family, thresh = 1e-04,
  nlam = 20, lambda = NULL, inner.iter = 100, outer.iter = 100,
  outer.thresh = 1e-04, optim.step.size = 0.8, reset = 10,
  min.frac = 0.05, verbose = FALSE)
}
\description{
Get the sequence of tuning parameters and then make the call to the C code that implements the adaptive group lasso
}

