\name{lagr.fit.inner}
\alias{lagr.fit.inner}
\title{Fit a LAGR model}
\usage{
lagr.fit.inner(x, y, coords, loc, family, varselect.method, tuning, predict,
  simulation, verbose, kernel.weights = NULL, prior.weights = NULL,
  longlat = FALSE, N = 1, oracle = oracle)
}
\description{
Fit a LAGR model
}

