% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lagr.fit.inner}
\alias{lagr.fit.inner}
\title{Fit a local model via the local adaptive group lasso}
\usage{
lagr.fit.inner(x, y, coords, loc, family, varselect.method, oracle, tuning,
  predict, simulation, verbose, kernel.weights = NULL, prior.weights = NULL,
  longlat = FALSE)
}
\arguments{
\item{x}{matrix of observed covariates}

\item{y}{vector of observed responses}

\item{loc}{location at which to fit a model}

\item{family}{exponential family distribution of the response}

\item{varselect.method}{criterion to minimize in the regularization step of
fitting local models - options are \code{AIC}, \code{AICc}, \code{BIC},
\code{GCV}}

\item{tuning}{logical indicating whether this model will be used to tune the
bandwidth, in which case only the tuning criteria are returned}

\item{kernel.weights}{vector of observation weights from the kernel}

\item{prior.weights}{vector of prior observation weights provided by the user}

\item{longlat}{\code{TRUE} indicates that the coordinates are specified in
  longitude/latitude, \code{FALSE} indicates Cartesian coordinates. Default
  is \code{FALSE}.}
}
\value{
list of coefficients, nonzero coefficient identities, and tuning data
}
\description{
This function augments the covariates with local interactions, drops
observations with zero weight, runs the adaptive grouped lasso, computes the
residuals, and returns the necessary values for tuning or reporting.
}

