% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lagr.fit.inner}
\alias{lagr.fit.inner}
\title{Fit a LAGR model}
\usage{
lagr.fit.inner(x, y, coords, loc, family, varselect.method, oracle, tuning,
  predict, simulation, verbose, kernel.weights = NULL, prior.weights = NULL,
  longlat = FALSE)
}
\description{
Fit a LAGR model
}

