% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lagr}
\alias{lagr}
\title{Fit a model via local, adaptive grouped regularization}
\usage{
lagr(formula, data, family = c("gaussian", "binomial", "poisson", "Cox"),
  weights = NULL, coords, fit.loc = NULL, tuning = FALSE,
  predict = FALSE, simulation = FALSE, oracle = NULL, kernel, bw = NULL,
  varselect.method = c("AIC", "BIC", "AICc"), verbose = FALSE, longlat,
  tol.loc = NULL, bw.type = c("dist", "knn", "nen"), D = NULL,
  resid.type = c("deviance", "pearson"), na.action = c(na.omit, na.fail,
  na.pass), contrasts = NULL)
}
\arguments{
\item{formula}{symbolic representation of the model}

\item{data}{data frame containing observations of all the terms represented in the formula}

\item{weights}{vector of prior observation weights (due to, e.g., overdispersion). Not related to the kernel weights.}

\item{family}{exponential family distribution of the response}

\item{coords}{matrix of locations, with each row giving the location at which the corresponding row of data was observed}

\item{fit.loc}{matrix of locations where the local models should be fitted}

\item{longlat}{\code{TRUE} indicates that the coordinates are specified in longitude/latitude, \code{FALSE} indicates Cartesian coordinates. Default is \code{FALSE}.}

\item{kernel}{kernel function for generating the local observation weights}

\item{bw}{bandwidth parameter}

\item{bw.type}{type of bandwidth - options are \code{dist} for distance (the default), \code{knn} for nearest neighbors (bandwidth a proportion of \code{n}), and \code{nen} for nearest effective neighbors (bandwidth a proportion of the sum of squared residuals from a global model)}

\item{tol.loc}{tolerance for the tuning of an adaptive bandwidth (e.g. \code{knn} or \code{nen})}

\item{varselect.method}{criterion to minimize in the regularization step of fitting local models - options are \code{AIC}, \code{AICc}, \code{BIC}, \code{GCV}}

\item{resid.type}{type of residual to use (relevant for non-gaussian response) - options are \code{deviance} and \code{pearson}}

\item{tuning}{logical indicating whether this model will be used to tune the bandwidth, in which case only the tuning criteria are returned}

\item{D}{pre-specified matrix of distances between locations}

\item{verbose}{print detailed information about our progress?}
}
\value{
list containing the local models.
}
\description{
\code{lagr} fits a LAGR model This method fits a local model at each location indicated by \code{fit.loc}.
}

