% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lagr}
\alias{lagr}
\title{Fit a lagr model}
\usage{
lagr(formula, data, family, weights = NULL, coords, fit.loc = NULL,
  tuning = FALSE, predict = FALSE, simulation = FALSE, oracle = NULL,
  kernel, bw = NULL, varselect.method = c("AIC", "BIC", "AICc"),
  verbose = FALSE, longlat, tol.loc = NULL, bw.type = c("dist", "knn",
  "nen"), D = NULL, resid.type = c("deviance", "pearson"))
}
\description{
Fit a lagr model
}

