% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lagr.r
\name{lagr}
\alias{lagr}
\title{Fit a model via local, adaptive grouped regularization}
\usage{
lagr(formula, data, family = gaussian(), weights = NULL, coords,
  fit.loc = NULL, tuning = FALSE, predict = FALSE, simulation = FALSE,
  oracle = NULL, kernel = NULL, bw = NULL, varselect.method = c("AIC",
  "BIC", "AICc", "wAIC", "wAICc"), verbose = FALSE, longlat = FALSE,
  tol.loc = NULL, bw.type = c("dist", "knn", "nen"), D = NULL,
  lambda.min.ratio = 0.001, n.lambda = 50, lagr.convergence.tol = 0.001,
  lagr.max.iter = 20, jacknife = FALSE, bootstrap.index = NULL,
  na.action = na.fail, contrasts = NULL)
}
\arguments{
\item{formula}{symbolic representation of the model}

\item{data}{data frame containing observations of all the terms represented in the formula}

\item{family}{exponential family distribution of the response}

\item{weights}{vector of prior observation weights (due to, e.g., overdispersion). Not related to the kernel weights.}

\item{coords}{matrix of locations, with each row giving the location at which the corresponding row of data was observed}

\item{fit.loc}{matrix of locations where the local models should be fitted}

\item{tuning}{logical indicating whether this model will be used to tune the bandwidth, in which case only the tuning criteria are returned}

\item{kernel}{kernel function for generating the local observation weights}

\item{bw}{bandwidth parameter}

\item{varselect.method}{criterion to minimize in the regularization step of fitting local models - options are \code{AIC}, \code{AICc}, \code{BIC}, \code{GCV}}

\item{verbose}{print detailed information about our progress?}

\item{longlat}{\code{TRUE} indicates that the coordinates are specified in longitude/latitude, \code{FALSE} indicates Cartesian coordinates. Default is \code{FALSE}.}

\item{tol.loc}{tolerance for the tuning of an adaptive bandwidth (e.g. \code{knn} or \code{nen})}

\item{bw.type}{type of bandwidth - options are \code{dist} for distance (the default), \code{knn} for nearest neighbors (bandwidth a proportion of \code{n}), and \code{nen} for nearest effective neighbors (bandwidth a proportion of the sum of squared residuals from a global model)}

\item{D}{pre-specified matrix of distances between locations}
}
\value{
list containing the local models.
}
\description{
\code{lagr} fits a LAGR model This method fits a local model at each location indicated by \code{fit.loc}.
}

