% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SGL}
\alias{SGL}
\title{Do the local adaptive grouped regularization}
\usage{
SGL(data, index, weights = NULL, type = "linear", maxit = 1000,
  thresh = 0.001, min.frac = 0.1, nlam = 20, delta = 2, gamma = 0.8,
  standardize = TRUE, verbose = FALSE, step = 1, reset = 10,
  alpha = 0.95, lambdas = NULL, adaptive = TRUE, unpenalized = NULL)
}
\description{
Do the local adaptive grouped regularization
}

